feat: add --json flag to `codex exec`
